hikvision_doorbell_sdk/requirements.txt
paho-mqtt>=2.0.0,<3.0
